--- test/unit_tests/braket/default_simulator/openqasm/test_circuit_builder.py
@@ -1,33 +1,0 @@
-import numpy as np
-from braket.devices import LocalSimulator
-from braket.ir.openqasm import Program
-
-
-def test_gphase():
-    qasm = """
-    qubit[2] qs;
-
-    int[8] two = 2;
-
-    gate x a { U(π, 0, π) a; }
-    gate cx c, a { ctrl @ x c, a; }
-    gate phase c, a {
-        gphase(π/2);
-        pow(1) @ ctrl(two) @ gphase(π) c, a;
-    }
-    gate h a { U(π/2, 0, π) a; }
-
-    h qs[0];
-    cx qs[0], qs[1];
-    phase qs[0], qs[1];
-
-    gphase(π);
-    inv @ gphase(π / 2);
-    negctrl @ ctrl @ gphase(2 * π) qs[0], qs[1];
-    
-    #pragma {"braket result amplitude '00', '01', '10', '11'";}
-    """
-    device = LocalSimulator("braket_oq3_sv")
-    result = device.run(Program(source=qasm)).result()
-    sv = [result.result_types[0].value[state] for state in ("00", "01", "10", "11")]
-    assert np.allclose(sv, [-1 / np.sqrt(2), 0, 0, 1 / np.sqrt(2)])
